// src/core/brightStars.generated.ts
export const NAVIGABLE_ADDITION_SEEDS: StarSeed[] = [
  { id: "keid", name: "Keid", ra: 1.113835, dec: -0.133568, distanceLy: 16.2573, pmRA: -2239.33, pmDec: -3419.86, rv: -42.7, spectralType: "K1V", luminosity: 0.3658, massSun: 0.7, appMag: 4.43, con: "Eri", bayer: "Omi-2", hip: 19849 },
  { id: "70-oph", name: "70 Oph", ra: 4.73619, dec: 0.043635, distanceLy: 16.5798, pmRA: 124.56, pmDec: -962.66, rv: -6.9, spectralType: "K0V SB", luminosity: 0.55, massSun: 0.7, appMag: 4.03, con: "Oph", hip: 88601 },
  { id: "altair", name: "Altair", ra: 5.195772, dec: 0.154781, distanceLy: 16.7302, pmRA: 536.82, pmDec: 385.54, rv: -26.1, spectralType: "A7IV-V", luminosity: 11.3763, massSun: 2, appMag: 0.76, con: "Aql", bayer: "Alp", hip: 97649 },
  { id: "guniibuu", name: "Guniibuu", ra: 4.517567, dec: -0.464307, distanceLy: 19.3518, pmRA: -473.69, pmDec: -1143.93, rv: 0.9, spectralType: "K2:III:", luminosity: 0.5681, massSun: 0.7, appMag: 4.33, con: "Oph", hip: 84405 },
  { id: "achird", name: "Achird", ra: 0.21422, dec: 1.009065, distanceLy: 19.4164, pmRA: 1087.11, pmDec: -559.65, rv: 8.2, spectralType: "G0V SB", luminosity: 1.2753, massSun: 1, appMag: 3.46, con: "Cas", bayer: "Eta", hip: 3821 },
  { id: "82-g-eri", name: "82 G. Eri", ra: 0.872301, dec: -0.75171, distanceLy: 19.7109, pmRA: 3038.08, pmDec: 726.34, rv: 86.7, spectralType: "G8V", luminosity: 0.6286, massSun: 1, appMag: 4.26, con: "Eri", hip: 15510 },
  { id: "del-pav", name: "δ Pav", ra: 5.27399, dec: -1.155095, distanceLy: 19.9229, pmRA: 1210.29, pmDec: -1130.34, rv: -21.7, spectralType: "G5IV-Vvar", luminosity: 1.2359, massSun: 1, appMag: 3.55, con: "Pav", bayer: "Del", hip: 99240 },
  { id: "bet-hyi", name: "β Hyi", ra: 0.112028, dec: -1.348341, distanceLy: 24.3274, pmRA: 2220.12, pmDec: 324.37, rv: 22.9, spectralType: "G2IV", luminosity: 3.6074, massSun: 1, appMag: 2.82, con: "Hyi", bayer: "Bet", hip: 2021 },
  { id: "vega", name: "Vega", ra: 4.873563, dec: 0.676903, distanceLy: 25.0446, pmRA: 201.02, pmDec: 287.46, rv: -12.1, spectralType: "A0Vvar", luminosity: 49.9344, massSun: 2, appMag: 0.03, con: "Lyr", bayer: "Alp", hip: 91262 },
  { id: "fomalhaut", name: "Fomalhaut", ra: 6.011133, dec: -0.517006, distanceLy: 25.1258, pmRA: 329.22, pmDec: -164.22, rv: 6.1, spectralType: "A3V", luminosity: 17.5873, massSun: 2, appMag: 1.17, con: "PsA", bayer: "Alp", hip: 113368 },
];